Handle creation of UDF collection groupings.
The API will now return sets of UDF collections fields that need to be
grouped together and sorted. For instance tree and plot stewardships
will have similar data and should be shown together on the tree detail.
EX:
plot - watering - 2014.07.10
tree - watering - 2014.07.09
plot - mulching - 2014.07.08
As a result we need to track the sort key for the fields (date in the
example above) and the add both fields to the model as a new cell type
that can handle this behavior.