Storyboard changes.
It's hard to tell what these actually do and I don't recall changing
anything on the storyboard in a meaningful way. The changes don't seem
to be causing any issues but I thought it would be better to keep the
separate in case we have any need to revert them.